Refactor: Used new API as only available API
After reviewing the connection logs for the last two years only one
person has accessed the old API. I've contacted them and they're ok with
me modifying the interface. Other users that have accessed it are bots
that use the example address specified in the old documetation. This
means that #15 will be closed without implementing the old API.